requirements.txt
google-generativeai==0.4.1
streamlit==1.32.2
python-dotenv==1.0.1